requirements.txt
flask
flask-socketio
eventlet
pycryptodome
